+++ rocclr/utils/flags.hpp
@@ -172,7 +172,7 @@ release(uint, HIP_LAUNCH_BLOCKING, 0,                                         \
 release(bool, PAL_ALWAYS_RESIDENT, false,                                     \
         "Force memory resources to become resident at allocation time")       \
 release(uint, HIP_HOST_COHERENT, 0,                                           \
-        "Coherent memory in hipHostMalloc, 0x1 = memory is coherent with host"\
+        "Coherent memory in hipExtHostAlloc, 0x1 = memory is coherent with host"\
         "0x0 = memory is not coherent between host and GPU")                  \
 release(uint, AMD_OPT_FLUSH, 1,                                               \
         "Kernel flush option , 0x0 = Use system-scope fence operations."      \